issue - 10

+++ contracts/libraries/leverageX/TradingStorageUtils.sol
@@ -115,7 +115,7 @@ library TradingStorageUtils {
         ITradingStorage.TradingStorage storage s = _getStorage();
         s.borrowingProvider = _borrowingProvider;
 
-        for (uint i = 0; i < s.lastCollateralIndex; ++i) {
+        for (uint8 i = 0; i < s.lastCollateralIndex; ++i) {
             updateCollateralApprove(i);
         }
